frameworks are stored in tmp directory before some of their architectures are removed, before we start tests they are copied back to ios-dependencies so we can use these architectures for testing

+++ src/main/resources/META-INF/copy-dependencies-back-to-target.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 
 APP_PATH="$1"
-echo "Dependencies will be copied back to target from tmp-target"
 cd "$APP_PATH"
-rm -rf ios-dependencies
-cp -a ../tmp-target ios-dependencies
-rm -rf ../tmp-target
+if [[ -d ../tmp-target ]]; then
+    echo "Dependencies will be copied back to target from tmp-target"
+    rm -rf ios-dependencies
+    cp -a ../tmp-target ios-dependencies
+    rm -rf ../tmp-target
+fi
